qfc: use doubles for intermediate values
We could lose precision when multiplying by integral_max with floats.
For doubles, we can count on up to 2^56 - 1 to be expressible somewhat
exactly.

+++ qfunctionconfigurator/functionconfig.h
@@ -20,7 +20,7 @@
 class Map {
 private:
     static constexpr int value_count = 5000;
-    using num = float;
+    using num = double;
     using integral = std::uint16_t;
     static constexpr integral integral_max = std::numeric_limits<integral>::max();
